Consolidate Types to GameSimulatorTypes
