--- src/pages/academy/gameSimulator/subcomponents/assetViewer/AssetViewerTypes.tsx
@@ -1,3 +1,0 @@
-export type AssetProps = {
-  assetPath: string;
-};
