--- src/pages/academy/gameSimulator/subcomponents/chapterPublisher/ChapterPublisherTypes.tsx
@@ -1,6 +1,0 @@
-import { ChapterDetail } from 'src/features/gameSimulator/GameSimulatorTypes';
-
-export type ChapterSimProps = {
-  chapterDetail: ChapterDetail;
-  chapterFilenames?: string[];
-};
