--- src/pages/academy/gameSimulator/subcomponents/chapterSimulator/ChapterSimulatorTypes.tsx
@@ -1,4 +1,0 @@
-export type Props = {
-  storageName: string;
-  s3TxtFiles: string[];
-};
